更改ScreenToCoordinateMapper的成员函数GetCoordinateLocation中类型转换错误的Bug，应将GetCoordinaX和GetCoordinateY的值转换为float类型

+++ NewUtilities/Mapper/ScreenToCoordinateMapper.cs
@@ -47,21 +47,8 @@ namespace Utilities.Mapper
         public double GetScreenY(double coordinateY) => YAxisMapper.MapToValue1(coordinateY);
         public double GetCoordinateX(double screenX) => XAxisMapper.MapToValue2(screenX);
         public double GetCoordinateY(double screenY) => YAxisMapper.MapToValue2(screenY);
-        public PointF GetScreenLocation(double coordinateX, double coordinateY) => new PointF((int)GetScreenX(coordinateX), (int)GetScreenY(coordinateY));
-        public PointF GetCoordinateLocation(double screenX, double screenY) => new PointF((int)GetCoordinateX(screenX), (int)GetCoordinateY(screenY));
-        //public void SetCoordinateXRange(double xLeft, double xRight)
-        //{
-        //    XAxisMapper.SetRange2(xLeft, xRight);
-        //    if(!isInitializing)
-        //        MapperStateChanged?.Invoke(this);
-        //}
-
-        //public void SetCoordinateYRange(double yTop, double yBottom)
-        //{
-        //    YAxisMapper.SetRange2(yTop, yBottom);
-        //    if (!isInitializing)
-        //        MapperStateChanged?.Invoke(this);
-        //}
+        public PointF GetScreenLocation(double coordinateX, double coordinateY) => new PointF((float)GetScreenX(coordinateX), (float)GetScreenY(coordinateY));
+        public PointF GetCoordinateLocation(double screenX, double screenY) => new PointF((float)GetCoordinateX(screenX), (float)GetCoordinateY(screenY));
 
         public void SetScreenArea(double left, double right, double top, double bottom)
         {
